--- fdb-record-layer-geophile/src/main/java/com/apple/foundationdb/record/geophile/SpatialFunctionNames.java
@@ -1,54 +1,0 @@
-/*
- * SpatialFunctionNames.java
- *
- * This source file is part of the FoundationDB open source project
- *
- * Copyright 2015-2019 Apple Inc. and the FoundationDB project authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.apple.foundationdb.record.geophile;
-
-import com.apple.foundationdb.annotation.API;
-
-/**
- * Key function names for (geo-)spatial indexes.
- */
-@API(API.Status.EXPERIMENTAL)
-public class SpatialFunctionNames {
-    /**
-     * Index a geospatial point.
-     *
-     * Arguments are double fields latitude, longitude.
-     */
-    public static final String GEO_POINT_Z = "geo_point_z";
-
-    /**
-     * Index a GeoJson string field.
-     */
-    public static final String GEO_JSON_Z = "geo_json_z";
-
-    /**
-     * Index a JTS geometry serialized as WKB to a byte string field.
-     */
-    public static final String GEO_WKB_Z = "geo_wkb_z";
-
-    /**
-     * Index a JTS geometry serialized as WKT to a string field.
-     */
-    public static final String GEO_WKT_Z = "geo_wkt_z";
-
-    private SpatialFunctionNames() {
-    }
-}
