--- fdb-record-layer-geophile/src/main/java/com/apple/foundationdb/record/geophile/SpatialIndexMaintainerFactory.java
@@ -1,61 +1,0 @@
-/*
- * SpatialIndexMaintainerFactory.java
- *
- * This source file is part of the FoundationDB open source project
- *
- * Copyright 2015-2019 Apple Inc. and the FoundationDB project authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.apple.foundationdb.record.geophile;
-
-import com.apple.foundationdb.annotation.API;
-import com.apple.foundationdb.record.metadata.Index;
-import com.apple.foundationdb.record.metadata.IndexValidator;
-import com.apple.foundationdb.record.provider.foundationdb.IndexMaintainer;
-import com.apple.foundationdb.record.provider.foundationdb.IndexMaintainerFactory;
-import com.apple.foundationdb.record.provider.foundationdb.IndexMaintainerState;
-import com.google.auto.service.AutoService;
-
-import javax.annotation.Nonnull;
-import java.util.Arrays;
-
-/**
- * A factory for {@link SpatialIndexMaintainer} indexes.
- */
-@AutoService(IndexMaintainerFactory.class)
-@API(API.Status.MAINTAINED)
-public class SpatialIndexMaintainerFactory implements IndexMaintainerFactory {
-    // TODO: Maybe SPATIAL, as well, with a different space.
-    static final String[] TYPES = { SpatialIndexTypes.GEOSPATIAL_GEOPHILE };
-
-    @Override
-    @Nonnull
-    public Iterable<String> getIndexTypes() {
-        return Arrays.asList(TYPES);
-    }
-
-    @Nonnull
-    @Override
-    public IndexValidator getIndexValidator(Index index) {
-        return new SpatialIndexValidator(index);
-    }
-
-    @Nonnull
-    @Override
-    public IndexMaintainer getIndexMaintainer(IndexMaintainerState state) {
-        return new SpatialIndexMaintainer(state);
-    }
-
-}
