--- fdb-record-layer-geophile/src/main/java/com/apple/foundationdb/record/geophile/package-info.java
@@ -1,24 +1,0 @@
-/*
- * package-info.java
- *
- * This source file is part of the FoundationDB open source project
- *
- * Copyright 2015-2019 Apple Inc. and the FoundationDB project authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Classes for doing geospatial queries with <a href="https://github.com/geophile/geophile">Geophile</a>.
- */
-package com.apple.foundationdb.record.geophile;
