--- fdb-record-layer-geophile/src/test/proto/test_records_geo.proto
@@ -1,51 +1,0 @@
-/*
- * test_records_geo.proto
- *
- * This source file is part of the FoundationDB open source project
- *
- * Copyright 2015-2019 Apple Inc. and the FoundationDB project authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-syntax = "proto2";
-
-package com.apple.foundationdb.record.geophile;
-
-option java_package = "com.apple.foundationdb.record.geophile";
-option java_outer_classname = "TestRecordsGeoProto";
-
-import "record_metadata_options.proto";
-
-message City {
-    message Location {
-        optional double latitude = 1;
-        optional double longitude = 2;
-    }
-    optional int32 geo_name_id = 1 [(field).primary_key = true];
-    optional string name = 2;
-    optional string name_ascii = 3;
-    optional string country = 4;
-    optional Location location = 5;
-}
-
-message Country {
-    optional int32 geo_name_id = 1 [(field).primary_key = true];
-    optional string name = 2;
-    optional string code = 3;
-    optional string shape = 4;
-}
-
-message RecordTypeUnion {
-    optional City _City = 1;
-    optional Country _Country = 2;
-}
