+++ fdb-record-layer-spatial/gradle.properties
