+++ fdb-record-layer-spatial/src/main/java/com/apple/foundationdb/record/spatial/geophile/GeophileCoveringPointRecord.java
@@ -1,0 +1,63 @@
+/*
+ * CoveringPointRecord.java
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2015-2019 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.apple.foundationdb.record.spatial.geophile;
+
+import com.apple.foundationdb.record.IndexEntry;
+import com.apple.foundationdb.tuple.Tuple;
+import com.geophile.z.SpatialObject;
+import com.geophile.z.spatialobject.d2.Point;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+/**
+ * Index entry with original point in the value portion.
+ *
+ * This allow more efficient filtering out of spatial join false positives.
+ */
+class GeophileCoveringPointRecord extends GeophileRecordImpl {
+    @Nonnull
+    private final SpatialObject spatialObject;
+
+    GeophileCoveringPointRecord(@Nonnull IndexEntry indexEntry, @Nullable Tuple prefix) {
+        super(indexEntry, prefix);
+        Tuple value = indexEntry.getValue();
+        this.spatialObject = new Point(value.getDouble(0), value.getDouble(1));
+    }
+
+    @Nonnull
+    @Override
+    public SpatialObject spatialObject() {
+        return spatialObject;
+    }
+
+    @Override
+    @SuppressWarnings("PMD.UselessOverridingMethod")
+    public boolean equals(Object o) {
+        return super.equals(o);
+    }
+
+    @Override
+    @SuppressWarnings("PMD.UselessOverridingMethod")
+    public int hashCode() {
+        return super.hashCode();
+    }
+}
