+++ fdb-record-layer-spatial/src/main/java/com/apple/foundationdb/record/spatial/geophile/GeophileSpatialFunctionNames.java
@@ -1,0 +1,54 @@
+/*
+ * SpatialFunctionNames.java
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2015-2019 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.apple.foundationdb.record.spatial.geophile;
+
+import com.apple.foundationdb.annotation.API;
+
+/**
+ * Key function names for Geophile indexes.
+ */
+@API(API.Status.EXPERIMENTAL)
+public class GeophileSpatialFunctionNames {
+    /**
+     * Index a geospatial point.
+     *
+     * Arguments are double fields latitude, longitude.
+     */
+    public static final String GEOPHILE_POINT_Z = "geophile_point_z";
+
+    /**
+     * Index a GeoJson string field.
+     */
+    public static final String GEOPHILE_JSON_Z = "geophile_json_z";
+
+    /**
+     * Index a JTS geometry serialized as WKB to a byte string field.
+     */
+    public static final String GEOPHILE_WKB_Z = "geophile_wkb_z";
+
+    /**
+     * Index a JTS geometry serialized as WKT to a string field.
+     */
+    public static final String GEOPHILE_WKT_Z = "geophile_wkt_z";
+
+    private GeophileSpatialFunctionNames() {
+    }
+}
